fixed date bug - I think
Fixed bug in calendar page where date was displayed incorrectly.  Moved the logic for checking if an event is the first of the day into its own utility function, and passed that array of ids down from the root calendar component, instead of doing the check later down the line.